update field verification in party controllers

+++ api/controllers/parties/deleteParty.js
@@ -11,7 +11,11 @@ const deleteParty = (req, res) => {
   const { id } = req.params;
 
   // Verify Fields
-  verifyFields(['id'], req.params, res);
+  const missingFields = verifyFields(['id'], req.body);
+
+  if (missingFields.length > 0) {
+    return res.status(422).json({ msg: `Fields missing: ${missingFields.join(', ')}` });
+  }
 
   Party.findOneAndRemove({ _id: id })
     .then(removedParty => {
